Put classes and constants in their own files separate from main
