+++ constants.py
@@ -1,0 +1,2 @@
+_INF = float('infinity')
+_SIZE = None
